Issue with player Inventory

+++ Assets/ScriptableObjects/Scripts/Enemy/BaseEnemy.cs
@@ -12,7 +12,7 @@ public abstract class BaseEnemy : MonoBehaviour
 
     protected float currentMoveSpeed;
     protected float currentHealth;
-    protected float currentDamage;
+    protected int currentDamage;
     protected bool canDamage;
 
     protected static int killedEnemy;
